Move name, tritonModelName, and effective_name to BaseModelProperties
Deduplicate shared fields and property that were identically defined in
HuggingFaceModelProperties, OpenCLIPModelProperties, and RandomModelProperties
by hoisting them into the common base class.
